Better addition and subtraction of zero
When adding, or subtracting, a 0-value to an Addition or Subtraction operator, and the operator has only one element, return that one element instead of a new operator. This is a small optimization that reduces the buildup of the huge operator trees that are often the result of complex operations.