Added support for spaceship of operator proxies
This is a bit of a hack. If two proxies are exactly the same, it converts everything to the units of the first operand of self and then does the comparison. I’m sure there’s a more optimal way to do this, but this works well enough for now.